Monthly Tariff Payment, PayDays And Bug Fixes

+++ src/main/java/com/bootcamp_proj/bootcampproj/additional_classes/BrtTransaction.java
@@ -63,7 +63,7 @@ public class BrtTransaction extends Transaction {
         }
     }
 
-    public int getCallLength() {
+    public double getCallLength() {
         return unixEnd - unixStart;
     }
 }
